Extract parse functions
