+++ src/language-css/parse/parse-media-query.js
@@ -1,0 +1,23 @@
+import postcssMediaQueryParser from "postcss-media-query-parser";
+import { addTypePrefix, addMissingType } from "./utils.js";
+
+const parse = postcssMediaQueryParser.default;
+
+function parseMediaQuery(params) {
+  let result = null;
+
+  try {
+    result = parse(params);
+  } catch {
+    // Ignore bad media queries
+    /* c8 ignore next 4 */
+    return {
+      type: "selector-unknown",
+      value: params,
+    };
+  }
+
+  return addTypePrefix(addMissingType(result), "media-");
+}
+
+export default parseMediaQuery;
