+++ src/language-css/parser-postcss.js
@@ -1,9 +1,6 @@
 import postcssParse from "postcss/lib/parse";
 import postcssLess from "postcss-less";
 import postcssScssParse from "postcss-scss/lib/scss-parse";
-import PostcssValuesParser from "postcss-values-parser/lib/parser.js";
-import PostcssSelectorParser from "postcss-selector-parser/dist/processor.js";
-import postcssMediaQueryParser from "postcss-media-query-parser";
 import createError from "../common/parser-create-error.js";
 import parseFrontMatter from "../utils/front-matter/parse.js";
 import { hasPragma } from "./pragma.js";
@@ -13,284 +10,12 @@ import {
   calculateLoc,
   replaceQuotesInInlineComments,
 } from "./loc.js";
-import hasSCSSInterpolation from "./utils/has-scss-interpolation.js";
-import hasStringOrFunction from "./utils/has-string-or-function.js";
 import isSCSSNestedPropertyNode from "./utils/is-scss-nested-property-node.js";
-import isSCSSVariable from "./utils/is-scss-variable.js";
-import stringifyNode from "./utils/stringify-node.js";
 import isModuleRuleName from "./utils/is-module-rule-name.js";
-
-const getHighestAncestor = (node) => {
-  while (node.parent) {
-    node = node.parent;
-  }
-  return node;
-};
-
-function parseValueNode(valueNode, options) {
-  const { nodes } = valueNode;
-  let parenGroup = {
-    open: null,
-    close: null,
-    groups: [],
-    type: "paren_group",
-  };
-  const parenGroupStack = [parenGroup];
-  const rootParenGroup = parenGroup;
-  let commaGroup = {
-    groups: [],
-    type: "comma_group",
-  };
-  const commaGroupStack = [commaGroup];
-
-  for (let i = 0; i < nodes.length; ++i) {
-    const node = nodes[i];
-
-    if (
-      options.parser === "scss" &&
-      node.type === "number" &&
-      node.unit === ".." &&
-      node.value.endsWith(".")
-    ) {
-      // Work around postcss bug parsing `50...` as `50.` with unit `..`
-      // Set the unit to `...` to "accidentally" have arbitrary arguments work in the same way that cases where the node already had a unit work.
-      // For example, 50px... is parsed as `50` with unit `px...` already by postcss-values-parser.
-      node.value = node.value.slice(0, -1);
-      node.unit = "...";
-    }
-
-    if (node.type === "func" && node.value === "selector") {
-      node.group.groups = [
-        parseSelector(
-          getHighestAncestor(valueNode).text.slice(
-            node.group.open.sourceIndex + 1,
-            node.group.close.sourceIndex
-          )
-        ),
-      ];
-    }
-
-    if (node.type === "func" && node.value === "url") {
-      const groups = node.group?.groups ?? [];
-
-      // Create a view with any top-level comma groups flattened.
-      let groupList = [];
-      for (let i = 0; i < groups.length; i++) {
-        const group = groups[i];
-        if (group.type === "comma_group") {
-          groupList = [...groupList, ...group.groups];
-        } else {
-          groupList.push(group);
-        }
-      }
-
-      // Stringify if the value parser can't handle the content.
-      if (
-        hasSCSSInterpolation(groupList) ||
-        (!hasStringOrFunction(groupList) &&
-          !isSCSSVariable(groupList[0], options))
-      ) {
-        const stringifiedContent = stringifyNode({
-          groups: node.group.groups,
-        });
-        node.group.groups = [stringifiedContent.trim()];
-      }
-    }
-    if (node.type === "paren" && node.value === "(") {
-      parenGroup = {
-        open: node,
-        close: null,
-        groups: [],
-        type: "paren_group",
-      };
-      parenGroupStack.push(parenGroup);
-
-      commaGroup = {
-        groups: [],
-        type: "comma_group",
-      };
-      commaGroupStack.push(commaGroup);
-    } else if (node.type === "paren" && node.value === ")") {
-      if (commaGroup.groups.length > 0) {
-        parenGroup.groups.push(commaGroup);
-      }
-      parenGroup.close = node;
-
-      /* c8 ignore next 3 */
-      if (commaGroupStack.length === 1) {
-        throw new Error("Unbalanced parenthesis");
-      }
-
-      commaGroupStack.pop();
-      commaGroup = commaGroupStack.at(-1);
-      commaGroup.groups.push(parenGroup);
-
-      parenGroupStack.pop();
-      parenGroup = parenGroupStack.at(-1);
-    } else if (node.type === "comma") {
-      parenGroup.groups.push(commaGroup);
-      commaGroup = {
-        groups: [],
-        type: "comma_group",
-      };
-      commaGroupStack[commaGroupStack.length - 1] = commaGroup;
-    } else {
-      commaGroup.groups.push(node);
-    }
-  }
-  if (commaGroup.groups.length > 0) {
-    parenGroup.groups.push(commaGroup);
-  }
-  return rootParenGroup;
-}
-
-function flattenGroups(node) {
-  if (
-    node.type === "paren_group" &&
-    !node.open &&
-    !node.close &&
-    node.groups.length === 1
-  ) {
-    return flattenGroups(node.groups[0]);
-  }
-
-  if (node.type === "comma_group" && node.groups.length === 1) {
-    return flattenGroups(node.groups[0]);
-  }
-
-  if (node.type === "paren_group" || node.type === "comma_group") {
-    return { ...node, groups: node.groups.map(flattenGroups) };
-  }
-
-  return node;
-}
-
-function addTypePrefix(node, prefix, skipPrefix) {
-  if (node && typeof node === "object") {
-    delete node.parent;
-    for (const key in node) {
-      addTypePrefix(node[key], prefix, skipPrefix);
-      if (
-        key === "type" &&
-        typeof node[key] === "string" &&
-        !node[key].startsWith(prefix) &&
-        (!skipPrefix || !skipPrefix.test(node[key]))
-      ) {
-        node[key] = prefix + node[key];
-      }
-    }
-  }
-  return node;
-}
-
-function addMissingType(node) {
-  if (node && typeof node === "object") {
-    delete node.parent;
-    for (const key in node) {
-      addMissingType(node[key]);
-    }
-    if (!Array.isArray(node) && node.value && !node.type) {
-      node.type = "unknown";
-    }
-  }
-  return node;
-}
-
-function parseNestedValue(node, options) {
-  if (node && typeof node === "object") {
-    for (const key in node) {
-      if (key !== "parent") {
-        parseNestedValue(node[key], options);
-        if (key === "nodes") {
-          node.group = flattenGroups(parseValueNode(node, options));
-          delete node[key];
-        }
-      }
-    }
-    delete node.parent;
-  }
-  return node;
-}
-
-function parseValue(value, options) {
-  // Inline javascript in Less
-  if (options.parser === "less" && value.startsWith("~`")) {
-    return { type: "value-unknown", value };
-  }
-
-  const valueParser = (source, options) =>
-    new PostcssValuesParser(source, options);
-
-  let result = null;
-
-  try {
-    result = valueParser(value, { loose: true }).parse();
-  } catch {
-    return {
-      type: "value-unknown",
-      value,
-    };
-  }
-
-  result.text = value;
-
-  const parsedResult = parseNestedValue(result, options);
-
-  return addTypePrefix(parsedResult, "value-", /^selector-/);
-}
-
-function parseSelector(selector) {
-  // If there's a comment inside of a selector, the parser tries to parse
-  // the content of the comment as selectors which turns it into complete
-  // garbage. Better to print the whole selector as-is and not try to parse
-  // and reformat it.
-  if (/\/\/|\/\*/.test(selector)) {
-    return {
-      type: "selector-unknown",
-      value: selector.trim(),
-    };
-  }
-
-  const selectorParser = (processor) => new PostcssSelectorParser(processor);
-
-  let result = null;
-
-  try {
-    selectorParser((result_) => {
-      result = result_;
-    }).process(selector);
-  } catch {
-    // Fail silently. It's better to print it as is than to try and parse it
-    // Note: A common failure is for SCSS nested properties. `background:
-    // none { color: red; }` is parsed as a NestedDeclaration by
-    // postcss-scss, while `background: { color: red; }` is parsed as a Rule
-    // with a selector ending with a colon. See:
-    // https://github.com/postcss/postcss-scss/issues/39
-    return {
-      type: "selector-unknown",
-      value: selector,
-    };
-  }
-
-  return addTypePrefix(result, "selector-");
-}
-
-function parseMediaQuery(params) {
-  let result = null;
-
-  try {
-    result = postcssMediaQueryParser.default(params);
-  } catch {
-    // Ignore bad media queries
-    /* c8 ignore next 4 */
-    return {
-      type: "selector-unknown",
-      value: params,
-    };
-  }
-
-  return addTypePrefix(addMissingType(result), "media-");
-}
+import parseValue from "./parse/parse-value.js";
+import parseSelector from "./parse/parse-selector.js";
+import parseMediaQuery from "./parse/parse-media-query.js";
+import { addTypePrefix } from "./parse/utils.js";
 
 const DEFAULT_SCSS_DIRECTIVE = /(\s*)(!default).*$/;
 const GLOBAL_SCSS_DIRECTIVE = /(\s*)(!global).*$/;
